requirements.txt
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-WebEngine==6.7.0
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
